update companion device auth code

+++ test/unittest/services/cpp/cases/cross_device_comm/cross_device_comm_manager_impl_test.cpp
@@ -94,7 +94,7 @@ public:
     }
 
 protected:
-    int32_t nextGlobalId_ = 1;
+    uint64_t nextGlobalId_ = 1;
     NiceMock<MockMiscManager> mockMiscManager_;
     NiceMock<MockUserIdManager> mockActiveUserIdManager_;
     std::shared_ptr<NiceMock<MockCrossDeviceChannel>> mockChannel_;
